fix: repaired the patient component
