+++ src/app/patient-search/patient-search.component.ts
@@ -1,9 +1,38 @@
-import { FormControl } from '@angular/forms';
 import { Component } from '@angular/core';
+import { OnInit } from '@angular/core';
+
+import { FormControl } from '@angular/forms';
+import { Observable } from 'rxjs';
+import { map, startWith } from 'rxjs/operators';
+export interface User {
+  name: string;
+}
 @Component({
   selector: 'app-patient-search',
   templateUrl: './patient-search.component.html',
 })
-export class PatientSearchComponent {
-  FormControl = new FormControl('')
+export class PatientSearchComponent implements OnInit {
+  myControl = new FormControl<string | User>('');
+  options: User[] = [{ name: 'Mary' }, { name: 'Shelley' }, { name: 'Igor' }, { name: 'Maria' }, { name: 'Mia' }, { name: 'John' }, { name: 'Sam' }, { name: 'Rose' }, { name: 'Mathew' }, { name: 'Patrick' }];
+  filteredOptions!: Observable<User[]>;
+
+  ngOnInit() {
+    this.filteredOptions = this.myControl.valueChanges.pipe(
+      startWith(''),
+      map(value => {
+        const name = typeof value === 'string' ? value : value?.name;
+        return name ? this._filter(name as string) : this.options.slice();
+      }),
+    );
+  }
+
+  displayFn(user: User): string {
+    return user && user.name ? user.name : '';
+  }
+
+  private _filter(name: string): User[] {
+    const filterValue = name.toLowerCase();
+
+    return this.options.filter(option => option.name.toLowerCase().includes(filterValue));
+  }
 }
